backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
openai
PyPDF2
python-docx
reportlab
python-multipart
python-docx
